% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_osmlanduse.R
\name{classify_osmlanduse}
\alias{classify_osmlanduse}
\title{Classify Landuse}
\usage{
classify_osmlanduse(osmlanduse, classes)
}
\arguments{
\item{osmlanduse}{Output of get_osmlanduse or measure_osmlanduse.}

\item{classes}{A table with the correspondence between tags and landuse classes. See examples}
}
\value{
An sf object adding columns representing landuse classes
}
\description{
Classify OpenStreetMap tags into predefined landuse classes
}
\examples{
landuse <- get_osmlanduse()
measures <- measure_osmlanduse(landuse)
data(clc)
classified <- classify_osmlanduse(measures,clc)
}
