% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osmlanduse.R
\name{get_osmlanduse}
\alias{get_osmlanduse}
\title{Landuse data}
\usage{
get_osmlanduse(area)
}
\arguments{
\item{area}{El área o región a descargar}
}
\value{
Un objeto sf (tabla)
}
\description{
Download data from OpenStreetMap útil para el mapa de usos del suelo.
}
\examples{
use <- get_osmlanduse("Partido de Luján")
}
