% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intafao}
\alias{intafao}
\title{INTA FAO Land Cover Classes}
\format{
A data frame with 65rows and 5 columns:
\describe{
  \item{class}{Number.}
  \item{class_number}{LCCS Class number.}
  \item{class_priority}{Priority to remove overlapping.}
  \item{class_name}{LCCS Class Name.}
  \item{osm_tag}{OpenStreetMap tag value associated with each class.}
}
}
\source{
Volante, J. N. 2009. Monitoreo de la Cobertura y el Uso del Suelo a partir de
sensores remotos. Instituto Nacional de Tecnología Agropecuaria,
}
\usage{
intafao
}
\description{
Legend harmonization between OSM tags and the National Institute of
Agricultural Technology (INTA) - FAO Land Use and Land
Cover classification system. Level one legend.
}
\details{
The data was adapted from Volante et al. (2009)
to be used in classification.
}
\keyword{datasets}
