% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_osmlanduse.R
\name{map_osmlanduse}
\alias{map_osmlanduse}
\title{Map Landuse}
\usage{
map_osmlanduse(classified, title = "osmlanduseR map")
}
\arguments{
\item{classified}{Output of classify_osmlanduse.}

\item{title}{Map title}
}
\value{
A tmap map
}
\description{
Map landuse classified
}
\examples{
landuse <- get_osmlanduse()
measures <- measure_osmlanduse(landuse)
data(clc)
classified <- classify_osmlanduse(measures,clc)
map <- map_osmlanduse(classified)
map
}
