% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_osmlanduse.R
\name{measure_osmlanduse}
\alias{measure_osmlanduse}
\title{Measure Landuse}
\usage{
measure_osmlanduse(osmlanduse, crs = 5347, units = "ha")
}
\arguments{
\item{osmlanduse}{Output of get_osmlanduse.}

\item{crs}{Set the CRS to measure area.}

\item{units}{The units for the area measures.}
}
\value{
An sf object with area measures
}
\description{
Measure area polygons of osm landuse removing overlapping
}
\examples{
landuse <- get_osmlanduse()
measures <- measure_osmlanduse(landuse)
}
