% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_osmlanduse.R
\name{classify_osmlanduse}
\alias{classify_osmlanduse}
\title{Classify Landuse}
\usage{
classify_osmlanduse(osmlanduse, osm_tag, class_name)
}
\arguments{
\item{osmlanduse}{Output of get_osmlanduse or measure_osmlanduse.}

\item{osm_tag}{A vector of OpenStreetMap tag values}

\item{class_name}{A vector of the same length of \code{osm_tag} assigning a class name or number to each
OpenStreetMap tag and the landuse classes.}
}
\value{
An sf object adding a column with the classified landuse classes.
}
\description{
Classify OpenStreetMap tags into predefined land use classes
}
\examples{
landuse <- get_osmlanduse()
measures <- measure_osmlanduse(landuse)
data(clc)
classified <- classify_osmlanduse(measures,clc)
}
