% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osmlanduse.R
\name{get_osmlanduse}
\alias{get_osmlanduse}
\title{Get OSM data for landuse/landcover analysis}
\usage{
get_osmlanduse(area = "Partido de Lujan", crop_to = "area")
}
\arguments{
\item{area}{An sf object or the place name of an area}

\item{crop_to}{Character string indicating if the result should be the
intersection with the polygon ("area", default) or the bounding box ("bbox").}

\item{crs}{Set the Coordinate reference system to transform the data.}
}
\value{
An sf object with landuse data.
}
\description{
Build an Overpass query to retrieve OSM data for land use and
land cover analysis and mapping.
}
\details{
The function retrieves OSM elements tagged with keys: natural, landuse, natural
amenity, aeroway, leisure, protected_area, waterway, highway and railway.
}
\examples{
lujan <- get_osmlanduse()
}
