% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_osmlanduse.R
\name{map_osmlanduse}
\alias{map_osmlanduse}
\title{Map Land use}
\usage{
map_osmlanduse(classified, title = "osmlanduseR map")
}
\arguments{
\item{classified}{Output of classify_osmlanduse.}

\item{title}{Map title}
}
\value{
A tmap map
}
\description{
Map land use classified
}
\examples{
area <-  "Lezica y Torrezuri, Partido de Luján"
lezica <- get_osmlanduse(area, crop_to = "bbox")
data(clc)
classified <- classify_osmlanduse(lezica$osmlanduse,osm_tag = clc$osm_tag,
class_name = clc$class_name)
map <- map_osmlanduse(classified)
map
}
